refactor: extract mutation_gate module + decompose dispatch_tool into 3 functions
- Extract MutationGateAllowance, MutationGateFailure, evaluate_mutation_gate,
  is_verifier_source_tool, is_refactor_gated_mutation_tool (~310 lines)
  from dispatch.rs into new mutation_gate.rs module
- Decompose dispatch_tool (305 lines) into:
  - validate_tool_access() — surface/namespace/tier/daemon access checks
  - build_success_response() — suggestions/budget/truncation/metrics/structured content
  - build_error_response() — error formatting with gate failure enrichment
  - dispatch_tool() — thin orchestrator (~140 lines)
